fix:Improves logging structure and clarity for proxy events
Refactors log messages to standardize event tagging and add
contextual metadata, making tracing and debugging easier.
Distinguishes between info, warn, debug, and error levels
more accurately, reducing false alarms and improving
operational observability. Updates user-facing error
messages for better clarity and localization.

+++ entity/src/user_service_apis.rs
@@ -25,7 +25,7 @@ pub struct Model {
     pub timeout_seconds: Option<i32>,
     pub max_request_per_min: Option<i32>,
     pub max_requests_per_day: Option<i32>,
-    pub max_tokens_per_day: Option<i32>,
+    pub max_tokens_per_day: Option<i64>,
     pub max_cost_per_day: Option<Decimal>,
     pub expires_at: Option<DateTime>,
     pub is_active: bool,
